backend/requirements.txt
fastapi
uvicorn
python-dotenv
cachetools
requests
openai
beautifulsoup4
webdriver-manager
selenium